Extract code into hook
